pyproject.toml
[tool.poetry]
name = "to-latex-converter"
version = "0.1.0"
description = "Convert PDF documents to LaTeX with math formula detection"
authors = [
    "Sardor Lutfullaev"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pdf2image = "^1.16.3"
omegaconf = ">=2.3.0,<3.0.0"
PyMuPDF = ">=1.23.0"
pytesseract = ">=0.3.10"
opencv-python = ">=4.8.0"
torch = ">=2.0.0"
transformers = ">=4.37.0"
fitz = ">=0.0.1.dev0"
pillow = ">=10.0.0"
pypdf = ">=3.11.0"
optimum = {extras = ["onnxruntime"], version = ">=1.12.0"}
pix2text = "^1.1.3.2"
jieba = "^0.42.1"
ipykernel = "^6.29.5"
polars = "^1.29.0"
datasets = "^3.6.0"
arxiv = "^2.2.0"
pypdf2 = "^3.0.1"
accelerate = "^1.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
